refactor(utils): move embodied actor helpers to nested_dict_process
Move the four nested-dict helpers (process_nested_dict_for_adv,
process_nested_dict_for_train, trim_nested_tensor_time_dim,
flatten_nested_tensor_time_batch) out of embodied_fsdp_actor_worker.py
into rlinf/utils/nested_dict_process.py, where the other nested-dict
utilities already live. EmbodiedFSDPActor now imports them from there;
fsdp_nft_policy_worker (the only other caller of
process_nested_dict_for_train) is updated accordingly. Addresses review
comment on PR #1478.
